+++ com.unity.render-pipelines.high-definition/Editor/Material/Decal/DisplacableRectHandles.cs
@@ -15,6 +15,7 @@ namespace UnityEditor.Rendering.HighDefinition
         Color m_MonochromeHandleColor;
         Color m_WireframeColor;
         Color m_WireframeColorBehind;
+        bool allowsNegative;
 
         /// <summary>The position of the center of the box in Handle.matrix space. On plane z=0.</summary>
         public Vector2 center { get; set; }
@@ -58,9 +59,10 @@ namespace UnityEditor.Rendering.HighDefinition
             }
         }
 
-        public DisplacableRectHandles(Color baseColor)
+        public DisplacableRectHandles(Color baseColor, bool allowsNegative = false)
         {
             this.baseColor = baseColor;
+            this.allowsNegative = allowsNegative;
         }
 
         /// <summary>Draw the rect.</summary>
@@ -258,7 +260,8 @@ namespace UnityEditor.Rendering.HighDefinition
                             case NamedEdge.Bottom: topPosition.y -= delta; break;
                         }
 
-                        EnsureEdgeFacesOutsideForSymetry(theChangedEdge, ref leftPosition, ref rightPosition, ref topPosition, ref bottomPosition);
+                        if (!allowsNegative)
+                            EnsureEdgeFacesOutsideForSymetry(theChangedEdge, ref leftPosition, ref rightPosition, ref topPosition, ref bottomPosition);
                     }
 
                     if (useHomothety)
@@ -278,13 +281,14 @@ namespace UnityEditor.Rendering.HighDefinition
                                 break;
                         }
 
-                        EnsureEdgeFacesOutsideForHomothety(theChangedEdge, ref leftPosition, ref rightPosition, ref topPosition, ref bottomPosition);
+                        if (!allowsNegative)
+                            EnsureEdgeFacesOutsideForHomothety(theChangedEdge, ref leftPosition, ref rightPosition, ref topPosition, ref bottomPosition);
                     }
 
                     var max = new Vector2(rightPosition.x, topPosition.y);
                     var min = new Vector2(leftPosition.x, bottomPosition.y);
 
-                    if (!useSymetry && !useHomothety)
+                    if (!useSymetry && !useHomothety && !allowsNegative)
                         EnsureEdgeFacesOutsideForOtherTransformation(ref max, ref min);
 
                     center = (max + min) * .5f;